+++ prettier.js
@@ -1,7 +1,5 @@
-module.exports = {
-  extends: [
-    // Turns off all eslint rules that are unnecessary or might conflict with Prettier,
-    // lints code which is formatted different than prettier rules
-    "plugin:prettier/recommended",
-  ]
-};
+const prettierConfigs = require("./configs/prettier.js");
+
+module.exports = [
+  ...prettierConfigs
+];